Move `discovery` to `virtool_workflow`.
