+++ virtool_workflow_runtime/config.py
@@ -3,8 +3,6 @@ from virtool_workflow.config.configuration import config_fixture
 REDIS_CONNECTION_STRING_ENV = "VT_REDIS_CONNECTION_STRING"
 REDIS_JOB_LIST_NAME_ENV = "VT_REDIS_JOB_LIST_NAME"
 NO_SENTRY_ENV = "VT_NO_SENTRY"
-MONGO_DATABASE_CONNECTION_STRING_ENV = "VT_DB_CONNECTION_STRING"
-MONGO_DATABASE_NAME_ENV = "VT_DB_NAME"
 
 
 @config_fixture(env=REDIS_CONNECTION_STRING_ENV, default="redis://localhost:6379")
@@ -22,16 +20,4 @@ def redis_job_list_name(_):
 @config_fixture(env=NO_SENTRY_ENV, default=True)
 def no_sentry(_):
     """A flag indicating whether or not to run sentry checks."""
-    ...
-
-
-@config_fixture(env=MONGO_DATABASE_NAME_ENV, default="virtool")
-def db_name(_):
-    """The MongoDB database name."""
-    ...
-
-
-@config_fixture(env=MONGO_DATABASE_CONNECTION_STRING_ENV, default="mongodb://localhost:27017")
-def db_connection_string(_):
-    """The MongoDB database connection string/url."""
     ...
