+++ virtool_workflow_runtime/fixture_loading.py
@@ -1,10 +1,12 @@
 from typing import Iterable
+
+from virtool_workflow import discovery
 from virtool_workflow.fixtures.scope import FixtureScope
-from virtool_workflow_runtime import discovery
 
 
 class InitializedWorkflowFixtureScope(FixtureScope):
     """A WorkflowFixtureScope that imports fixtures from a set of modules upon `__enter__`."""
+
     def __init__(self, fixture_plugins: Iterable[str], *providers, **instances):
         self.fixture_plugins = fixture_plugins
         super().__init__(*providers, **instances)
@@ -12,4 +14,3 @@ class InitializedWorkflowFixtureScope(FixtureScope):
     def __enter__(self):
         discovery.load_fixture_plugins(self.fixture_plugins)
         return super(InitializedWorkflowFixtureScope, self).__enter__()
-
